// ts/types.ts
export type HexString = `0x${string}`
// ...
export type TToken = {
  moonpay_symbol?: string
  token_symbol: string;
  coin_symbol: string;
  name: string;
  address: `0x${string}`;
  default_decimals: number;
  coingecko_token_id: string;
  coin_main_hex_color: string;
  about: string;
  website: string;
  whitepaper: string;
  content: string;
  twitter: string;
  reddit: string;
  telegram: string;
  blockchain_explorer: string;
  coingecko: string;
  messari: string;
  erc20_address?: string;
  logoURI: string;
  priceUSD: number;
  chainId?: number;
  amount: string;
};